pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.133.1",
    "fastmcp>=3.0.2",
    "langchain>=1.2.4",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]
